# openapi_spec.yaml
openapi: 3.0.3
info:
  title: QuickBooks Sales Reporter API
  description: |
    API para consultar datos de ventas de QuickBooks Online con cache automático.
    
    Este servidor proporciona acceso a datos de ventas mensuales y anuales desde QuickBooks Online,
    con un sistema de cache inteligente que mantiene los datos actualizados automáticamente.
    
    **Características principales:**
    - 📊 Consultas de ventas mensuales con detalles de facturas y recibos
    - 📈 Reportes anuales con análisis de tendencias y mejores/peores meses
    - 🔄 Cache automático con actualizaciones programadas
    - 📅 Datos históricos disponibles para múltiples períodos
    - 🎯 API REST estándar compatible con OpenWebUI
    
    **Casos de uso comunes:**
    - "¿Cómo van las ventas de este mes?"
    - "¿Cuál fue el mejor mes del año?"
    - "Muéstrame las ventas anuales de 2024"
    - "¿Cuántas facturas se emitieron en julio?"
    
    Desarrollado por KH LLOREDA, S.A. para análisis de ventas empresariales.
  version: "2.0.0"
  contact:
    name: KH LLOREDA, S.A.
    email: lopd@khlloreda.com
    url: https://github.com/jordiportal/Quickbooks
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Servidor de desarrollo local
  - url: http://localhost:8000
    description: Servidor Docker local
  - url: https://quickbooks-sales.yourdomain.com
    description: Servidor de producción
...
paths:
  /api/sales:
    get:
      tags: [sales]
      summary: Obtener ventas del mes actual
      description: |
        Retorna los datos de ventas del mes actual con detalles de facturas y recibos.
        Los datos se obtienen del cache cuando están disponibles, o directamente de QuickBooks si es necesario.
      operationId: getCurrentMonthSales
      responses:
        '200':
          description: Datos de ventas obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthlySalesResponse'
              examples:
                current_month:
                  summary: Ventas del mes actual
                  value:
                    período: "08/2025"
                    fecha_inicio: "2025-08-01"
                    fecha_fin: "2025-08-31"
                    total_ventas: 15234.50
                    recibos_de_venta:
                      cantidad: 45
                      total: 8234.30
                    facturas:
                      cantidad: 32
                      total: 7000.20
                    from_cache: true
                    last_updated: "2025-08-06T15:30:00"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/sales/{year}/{month}:
    get:
      tags: [sales]
      summary: Obtener ventas de un mes específico
      description: |
        Retorna los datos de ventas para un año y mes específicos.
        Útil para consultar datos históricos o comparar períodos.
      operationId: getSpecificMonthSales
      parameters:
        - name: year
          in: path
          required: true
          description: Año de consulta (ej. 2024)
          schema:
            type: integer
            minimum: 2020
            maximum: 2030
            example: 2024
        - name: month
          in: path
          required: true
          description: Mes de consulta (1-12)
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 7
      responses:
        '200':
          description: Datos de ventas del período específico
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthlySalesResponse'
        '400':
          description: Parámetros de fecha inválidos
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No se encontraron datos para el período especificado
        '500':
          $ref: '#/components/responses/InternalError'
  /api/annual:
    get:
      tags: [annual]
      summary: Obtener reporte anual del año actual
      description: |
        Retorna un reporte completo de ventas anuales con desglose mensual,
        análisis de tendencias, mejor/peor mes, y estadísticas agregadas.
      operationId: getCurrentYearReport
      responses:
        '200':
          description: Reporte anual obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnualSalesResponse'
              examples:
                annual_report:
                  summary: Reporte anual 2025
                  value:
                    año: 2025
                    total_anual: 125486.75
                    meses:
                      "01":
                        nombre: "Enero"
                        data:
                          total_ventas: 8234.50
                          recibos_de_venta:
                            cantidad: 15
                            total: 4234.30
                          facturas:
                            cantidad: 12
                            total: 4000.20
                      "02":
                        nombre: "Febrero"
                        data:
                          total_ventas: 12456.25
                          recibos_de_venta:
                            cantidad: 23
                            total: 7456.15
                          facturas:
                            cantidad: 18
                            total: 5000.10
                    resumen:
                      promedio_mensual: 10457.23
                      mejor_mes:
                        mes: "Febrero"
                        ventas: 12456.25
                        periodo: "02/2025"
                      peor_mes:
                        mes: "Enero"
                        ventas: 8234.50
                        periodo: "01/2025"
                      meses_con_ventas: 8
                    from_cache: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/annual/{year}:
    get:
      tags: [annual]
      summary: Obtener reporte anual de un año específico
      description: |
        Retorna un reporte completo de ventas para un año específico.
        Incluye análisis comparativo y datos históricos.
      operationId: getSpecificYearReport
      parameters:
        - name: year
          in: path
          required: true
          description: Año para el reporte anual
          schema:
            type: integer
            minimum: 2020
            maximum: 2030
            example: 2024
      responses:
        '200':
          description: Reporte anual del año específico
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnualSalesResponse'
        '400':
          description: Año inválido
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No se encontraron datos para el año especificado
        '500':
          $ref: '#/components/responses/InternalError'
  /api/quarterly/{year}:
    get:
      tags: [annual]
      summary: Obtener reporte trimestral
      description: |
        Retorna datos de ventas agrupados por trimestres para un año específico.
        Útil para análisis de tendencias estacionales.
      operationId: getQuarterlyReport
      parameters:
        - name: year
          in: path
          required: true
          description: Año para el reporte trimestral
          schema:
            type: integer
            minimum: 2020
            maximum: 2030
            example: 2024
      responses:
        '200':
          description: Reporte trimestral obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuarterlySalesResponse'
        '400':
          description: Año inválido
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/comparison/{year1}/{year2}:
    get:
      tags: [annual]
      summary: Comparar ventas entre dos años
      description: |
        Compara los datos de ventas entre dos años y calcula diferencias,
        porcentajes de crecimiento/decrecimiento y tendencias.
      operationId: compareYears
      parameters:
        - name: year1
          in: path
          required: true
          description: Primer año para comparación
          schema:
            type: integer
            example: 2024
        - name: year2
          in: path
          required: true
          description: Segundo año para comparación
          schema:
            type: integer
            example: 2023
      responses:
        '200':
          description: Comparación entre años completada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YearComparisonResponse'
        '400':
          description: Años inválidos para comparación
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/cache/stats:
    get:
      tags: [admin]
      summary: Estadísticas del sistema de cache
      description: |
        Retorna estadísticas detalladas del sistema de cache incluyendo
        número de entradas, actualizaciones exitosas/fallidas, y estado general.
      operationId: getCacheStats
      responses:
        '200':
          description: Estadísticas del cache obtenidas exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/scheduler/status:
    get:
      tags: [admin]
      summary: Estado del scheduler automático
      description: |
        Retorna el estado actual del scheduler de actualizaciones automáticas,
        incluyendo jobs programados y empresas activas.
      operationId: getSchedulerStatus
      responses:
        '200':
          description: Estado del scheduler obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/force-update:
    post:
      tags: [admin]
      summary: Forzar actualización inmediata de datos
      description: |
        Ejecuta una actualización inmediata de los datos de ventas,
        omitiendo el cache y obteniendo información fresca de QuickBooks.
      operationId: forceUpdate
      responses:
        '200':
          description: Actualización completada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/force-annual-update:
    post:
      tags: [admin]
      summary: Forzar actualización anual completa
      description: |
        Ejecuta una actualización completa de todos los datos anuales,
        procesando todos los meses del año actual desde QuickBooks.
      operationId: forceAnnualUpdate
      responses:
        '200':
          description: Actualización anual completada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnualUpdateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  schemas:
    MonthlySalesResponse:
      type: object
      description: Respuesta con datos de ventas mensuales
      properties:
        período:
          type: string
          description: Período en formato MM/YYYY
          example: "08/2025"
        fecha_inicio:
          type: string
          format: date
          description: Fecha de inicio del período
          example: "2025-08-01"
        fecha_fin:
          type: string
          format: date
          description: Fecha de fin del período
          example: "2025-08-31"
        total_ventas:
          type: number
          format: float
          description: Total de ventas del período
          example: 15234.50
        recibos_de_venta:
          $ref: '#/components/schemas/TransactionSummary'
        facturas:
          $ref: '#/components/schemas/TransactionSummary'
        from_cache:
          type: boolean
          description: Indica si los datos provienen del cache
          example: true
        last_updated:
          type: string
          format: date-time
          description: Última actualización de los datos
          example: "2025-08-06T15:30:00"
        current_year:
          type: integer
          description: Año actual para navegación
          example: 2025
        current_month:
          type: integer
          description: Mes actual para navegación
          example: 8
    AnnualSalesResponse:
      type: object
      description: Respuesta con reporte anual completo
      properties:
        año:
          type: integer
          description: Año del reporte
          example: 2025
        total_anual:
          type: number
          format: float
          description: Total de ventas del año
          example: 125486.75
        meses:
          type: object
          description: Datos de ventas por mes
          additionalProperties:
            $ref: '#/components/schemas/MonthData'
        resumen:
          $ref: '#/components/schemas/AnnualSummary'
        from_cache:
          type: boolean
          description: Indica si los datos provienen del cache
          example: true
        current_year:
          type: integer
          description: Año actual para navegación
          example: 2025
    QuarterlySalesResponse:
      type: object
      description: Respuesta con datos trimestrales
      properties:
        año:
          type: integer
          description: Año del reporte
          example: 2024
        total_anual:
          type: number
          format: float
          description: Total anual calculado
          example: 98765.43
        trimestres:
          type: object
          description: Datos por trimestre
          additionalProperties:
            $ref: '#/components/schemas/QuarterData'
    YearComparisonResponse:
      type: object
      description: Comparación entre dos años
      properties:
        año_actual:
          type: integer
          example: 2024
        año_anterior:
          type: integer
          example: 2023
        ventas_actual:
          type: number
          format: float
          example: 125486.75
        ventas_anterior:
          type: number
          format: float
          example: 98765.43
        diferencia:
          type: number
          format: float
          description: Diferencia en ventas
          example: 26721.32
        porcentaje_cambio:
          type: number
          format: float
          description: Porcentaje de cambio
          example: 27.05
        tendencia:
          type: string
          enum: [crecimiento, decrecimiento, estable]
          example: "crecimiento"
...
    SchedulerStatusResponse:
      type: object
      description: Estado del scheduler de actualizaciones
      properties:
        scheduler_running:
          type: boolean
          description: Indica si el scheduler está activo
          example: true
        active_companies:
          type: integer
          description: Número de empresas registradas
          example: 1
        jobs:
          type: array
          description: Lista de jobs programados
          items:
            $ref: '#/components/schemas/ScheduledJob'
        companies:
          type: array
          description: Lista de IDs de empresas activas
          items:
            type: string
          example: ["9341455133679605"]
...
  responses:
    Unauthorized:
      description: No autenticado - Se requiere autorización con QuickBooks
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "No autenticado"
            code: "AUTH_REQUIRED"
    InternalError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Error interno del servidor"
            code: "INTERNAL_ERROR"
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        Autenticación basada en sesiones de Flask.
        Se establece automáticamente después de la autorización OAuth con QuickBooks.
security:
  - SessionAuth: []

# Metadatos adicionales para herramientas OpenAPI
...